move disk check to check mod
