+++ crates/pyo3/src/system.rs
@@ -9,8 +9,6 @@
 use pyo3::prelude::*;
 use pyo3::{exceptions, PyResult};
 use similar::{ChangeTag, TextDiff};
-use std::sync::mpsc;
-use std::thread;
 
 use fapolicy_analyzer::events;
 use fapolicy_analyzer::events::db::DB as EventDB;
@@ -18,21 +16,21 @@ use fapolicy_app::app::State;
 use fapolicy_app::cfg;
 use fapolicy_app::sys::deploy_app_state;
 use fapolicy_trust::stat::Status::*;
-use fapolicy_trust::stat::{check, Status};
 
-use super::trust::PyTrust;
 use crate::acl::{PyGroup, PyUser};
 use crate::analysis::PyEventLog;
 use crate::rules::PyRule;
 use crate::trust;
 use crate::{daemon, rules};
 
+use super::trust::PyTrust;
+
 #[pyclass(module = "app", name = "System")]
 #[derive(Clone)]
 /// An immutable view of host system state.
 /// This only a container for state, it has to be applied to the host system.
 pub struct PySystem {
-    rs: State,
+    pub(crate) rs: State,
 }
 impl From<State> for PySystem {
     fn from(rs: State) -> Self {
@@ -43,11 +41,6 @@ impl From<PySystem> for State {
     fn from(py: PySystem) -> Self {
         py.rs
     }
-}
-
-enum Update {
-    Items(Vec<Status>),
-    Done,
 }
 
 #[pymethods]
@@ -187,81 +180,8 @@ fn rules_difference(lhs: &PySystem, rhs: &PySystem) -> String {
     diff_lines.join("")
 }
 
-#[pyfunction]
-fn check_disk_trust(db: &PySystem, update: PyObject, done: PyObject) -> PyResult<()> {
-    let recs: Vec<_> = db
-        .rs
-        .trust_db
-        .values()
-        .into_iter()
-        .map(|r| r.clone())
-        .collect();
-
-    let sz = recs.len() / 100;
-    let chunks = recs.chunks(sz + 1);
-
-    let (tx, rx) = mpsc::channel();
-    let mut handles = vec![];
-
-    for chunk in chunks {
-        let ttx = tx.clone();
-        let recs = chunk.to_vec();
-        let t = thread::spawn(move || {
-            let updates = recs
-                .into_iter()
-                .flat_map(|r| check(&r.trusted))
-                .collect::<Vec<_>>();
-            if ttx.send(Update::Items(updates)).is_err() {
-                eprintln!("failed to send Items msg");
-            };
-        });
-        handles.push(t);
-    }
-
-    let ttx = tx.clone();
-    thread::spawn(move || {
-        for handle in handles {
-            if handle.join().is_err() {
-                eprintln!("failed to join update handle");
-            };
-        }
-        if ttx.send(Update::Done).is_err() {
-            eprintln!("failed to send Done msg");
-        };
-    });
-
-    thread::spawn(move || {
-        let mut cnt = 0;
-        loop {
-            if let Ok(u) = rx.recv() {
-                match u {
-                    Update::Items(i) => {
-                        cnt += 1;
-                        let r: Vec<_> = i.into_iter().map(PyTrust::from).collect();
-                        Python::with_gil(|py| {
-                            if update.call1(py, (r, cnt)).is_err() {
-                                eprintln!("failed make 'update' callback");
-                            }
-                        });
-                    }
-                    Update::Done => break,
-                };
-            }
-        }
-
-        Python::with_gil(|py| {
-            if done.call0(py).is_err() {
-                eprintln!("failed to make 'done' callback");
-            }
-        });
-    });
-
-    Ok(())
-}
-
 pub fn init_module(_py: Python, m: &PyModule) -> PyResult<()> {
     m.add_class::<PySystem>()?;
     m.add_function(wrap_pyfunction!(rules_difference, m)?)?;
-    m.add_function(wrap_pyfunction!(check_disk_trust, m)?)?;
     Ok(())
 }
